created files: requirements.txt, Dockerfile, dependencies.py added a base logic for views in app.py and for crud.py fixed some issues

+++ app/schema.py
@@ -1,5 +1,4 @@
 import datetime
-from decimal import Decimal
 
 from typing import Literal
 
@@ -27,11 +26,11 @@ class CreateAdv(BaseModel):
     title: str
     author: str
     description: str
-    price: Decimal
+    price: float
 
 
 class UpdateAdv(BaseModel):
     title: str | None = None
     author: str | None = None
     description: str | None = None
-    price: Decimal | None = None
+    price: float | None = None
